Support registering class methods in jails
A new `allow_class_method` Jail call is used to whitelist access to
class methods. The same Jail is now returned for both instances of a
class and the class itself, but the allowed class method list is checked
instead for class methods.

No access to constants apart from String is given, but this tightens
access to the .new method on all other classes by restricting it to
String. Access to other classes would have to be via return values or
by being set in variables.