Move to Ruby 2.7+ (#45)
* Use Ruby syntax highlighting in README

* Remove redundant patch file

Since ruby_parser 2.0.0 it stopped using StringIO and now uses
StringScanner. That means the patch hasn't applied since 2008.

Fixes: eb79570c2794415cc3d5d960cccb7ccda08f136f

* Remove unused rubyparser_bug.rb

Fixes: eb79570c2794 ("Initial commit")

* Remove trailing whitespace

* Move to Ruby 2.7+

Ruby 2.5 and 2.6 are unsupported and providing Ruby 3 compatibility is
easier with Ruby 2.7 as the lower bound.

* Reduce version compatibility to Ruby < 3

Ruby 3 has changed keyword argument handling and safemode doesn't deal
with that correctly. That means it's broken in various cases.