Drop Ruby < 2.7 and deal with kwargs (#44)
* Copy process_call implementation from ruby2ruby

This brings it closer to the original, making it easier to see what's
going on. Looking at the git log of ruby2ruby there were no
modifications in this area. The only difference in process_call is that
process_call_receiver is jailed. This implementation actually supports
kwargs and kwsplat.

It also adds tests for kwargs and kwsplat.

* Deal with kwargs

Ruby 3 has started to separate args and kwargs.

* Add context for kwargs in tests

* Run tests on Ruby 3.0

---------

Co-authored-by: Oleh Fedorenko <ofedoren@redhat.com>